+++ sample-app/src/main/java/rx/android/samples/ListenInOutActivity.java
@@ -12,7 +12,7 @@ import rx.Observer;
 import rx.Subscription;
 import rx.observables.ConnectableObservable;
 
-import static rx.android.content.ContentObservable.bindActivity;
+import static rx.android.app.AppObservable.bindActivity;
 
 /**
  * Activity that binds to a counting sequence and is able to listen in and out to that
